feat(connected-and-connect-wallet): move to general
