+++ src/components/General/ConnectedWallet.tsx
@@ -4,7 +4,7 @@ import { useConnectWallet } from '@web3-onboard/react'
 import { useEffect, useState } from 'react'
 import { ethers } from 'ethers'
 
-import { formatToFixed4, getFormattedWalletAddress } from '../../../utilities'
+import { formatToFixed4, getFormattedWalletAddress } from '../../utilities'
 
 const ConnectedWallet = () => {
   const [{ wallet }] = useConnectWallet()
